+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 > This project follows the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) format.
 
 ## [Unreleased]
+* fix race condition when computing values.
 
 ## [0.4.1] - 2026-03-09
 * Move documentation to packages readme.